+++ src/hooks/useTheme.ts
@@ -1,30 +1,5 @@
 import { useState } from "react";
-
-export type ThemeName = "light" | "dark";
-
-export interface Theme {
-  background: string;
-  text: string;
-  componentBg: string;
-  border: string;
-  buttonBg: string;
-  buttonText: string;
-  selectedItemBg: string;
-  selectedItemText: string;
-  cardShadow: string;
-  hoverBg: string;
-  inputBg: string;
-  statusColors: {
-    "To Do": string;
-    "In Progress": string;
-    Done: string;
-  };
-  error: {
-    background: string;
-    color: string;
-    border: string;
-  };
-}
+import type { Theme, ThemeName } from "../types";
 
 const themes: Record<ThemeName, Theme> = {
   light: {
